requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── 3D Vision ─────────────────────────────────────────────────────────────────
open3d>=0.18.0
numpy>=1.24.0,<2.0.0

# ── Graph Neural Networks ─────────────────────────────────────────────────────
torch-geometric>=2.4.0
# Install PyG dependencies separately for your CUDA version:
# pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv
#   -f https://data.pyg.org/whl/torch-2.1.0+cu118.html

# ── NLP / LLM Grounding ───────────────────────────────────────────────────────
transformers>=4.40.0
tokenizers>=0.19.0
accelerate>=0.29.0
sentence-transformers>=2.7.0

# ── CLIP (for zero-shot object-language matching) ─────────────────────────────
openai-clip @ git+https://github.com/openai/CLIP.git
Pillow>=10.0.0

# ── Simulation ────────────────────────────────────────────────────────────────
pybullet>=3.2.6
gymnasium>=0.29.1

# ── Reinforcement Learning ────────────────────────────────────────────────────
stable-baselines3>=2.2.1
sb3-contrib>=2.2.1
tensorboard>=2.15.0

# ── Scene Graph / Graph utils ─────────────────────────────────────────────────
networkx>=3.2.0
scipy>=1.11.0

# ── ONNX / Production Inference ───────────────────────────────────────────────
onnx>=1.15.0
onnxruntime>=1.17.0
# For GPU: onnxruntime-gpu>=1.17.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client>=0.20.0

# ── Data / Serialization ──────────────────────────────────────────────────────
pyyaml>=6.0.1
msgpack>=1.0.7
h5py>=3.10.0

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
plotly>=5.18.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.2.0

# ── Optional: TensorRT (Linux + CUDA only) ────────────────────────────────────
# tensorrt>=8.6.0
# pycuda>=2023.1
